Add React payment page frontend, use Stripe Checkout (#16)
This uses the new Stripe Checkout API, which provides a payment page that supports Apple Pay page and is SCA compliant. Even though it's on Stripe's domain, it should be a good trade-off, as we'll get all future payment methods (SEPA, EU Direct Debit etc) for free when they're support in Checkout later this year. An example of this Checkout page in action can be seen deployed at [Our World in Data donations page](https://ourworldindata.org/donate).

The main donation page is rebuilt with React and hooks to manage the validation state.

The `/support` URL prefix is replaced with `/sponsor` to mimic what GitHub uses for their sponsorship page.